requirements.txt
Flask==2.1.1
Flask-SocketIO==5.1.1
Werkzeug==2.0.3
openai
python-dotenv==1.0.0
requests==2.28.2
